Improved async and sync in open ai and moved prompt to chat route
